+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `textfield`
   - inconsistencies of autovalidation when it's turned off.
 - Tapping on tappable components no longer triggers tap highlight
+- `checkbox`
+  - Changed display to `inline-flex` to fix layout issues
 
 ### Changed